refactor: Move taint and drain and set and check component commands to their own packages to be more explicit
